Implement Some UI Changes
