+++ src/components/cards/class_cards/index.scss
