+++ src/components/cards/hiring_cards/index.scss
