+++ src/components/modals/add_leaders/index.scss
@@ -1,12 +1,12 @@
-@use '../../styles/variables';
-@use '../../styles/mixins';
+@use '../../../styles/variables';
+@use '../../../styles/mixins';
 
 .add-leaders-modal{
     width: 25vw;
     height: 30vh;
     position: absolute;
-    top: 50%;
-    left: 50%;
+    top: 40%;
+    left: 83.5%;
     transform: translate(-50%, -50%);
     background-color: variables.$wild-white;
     @include mixins.box-shadow(0, 8, 50, -5, variables.$wild-grey);
