+++ src/components/modals/create_team/index.scss
@@ -1,5 +1,5 @@
-@use '../../styles/variables';
-@use '../../styles/mixins';
+@use '../../../styles/variables';
+@use '../../../styles/mixins';
 
 .hiring-post-modal{
     width: 50vw;
