+++ src/screens/class_management/view_class/index.js
@@ -43,9 +43,9 @@ function ViewClass() {
     },
     {
       id: 3,
-      label: 'Team Leaders',
-      className: 'team-leaders',
-      path: `/classes/${classId}/teamleaders`,
+      label: 'Teams',
+      className: 'teams',
+      path: `/classes/${classId}/teams`,
     },
     {
       id: 4,
