+++ src/screens/team_management/teams/index.scss
@@ -1,0 +1,6 @@
+@use '../../../styles/variables';
+@use '../../../styles/mixins';
+
+.table-body{
+    height: 100%;
+}
